refactor(ui,agent): dedupe session code — reuse canonical helpers, kill copy-paste
Deepscan cleanup of code authored this session, against the "no >1 function
with the same purpose / no AI slop / general-not-narrow" bar:

- chat-augmentation: drop local parseJsonObjectFromModelText; use core's
  canonical parseJSONObjectFromText (JSON5 + fenced-block tolerant, returns
  null for arrays) — fixes silent drops of fenced/loose model output on the
  cerebras-direct / local-gte-small hosts this recovery path targets.
- chat-augmentation: dedup recovery queries BEFORE slicing so "up to 3
  distinct queries" is honored when the model emits leading duplicates.
- useChatSend: extract the 6x copy-pasted empty-assistant-placeholder filter
  into one dropEmptyAssistantPlaceholder closure (no lockstep drift).
- transports: hoist headersToRecord / methodAllowsBody / bodyToString /
  isStreamingRequest into the shared ./transport.ts; native-cloud, android and
  desktop import them instead of keeping byte-identical copies. Tightens
  native-cloud's body normalizer from `unknown` to `string | undefined`.

biome clean; ui+agent typecheck clean (only pre-existing unbuilt-dep errors);
useChatSend 7/7 + transport suites 24/24 pass.
